updated composer and tests; refactor in storage adapters
