+++ src/Storage/Adapter/Libmemcached.php
@@ -105,24 +105,6 @@ class Libmemcached extends AbstractAdapter
     }
 
     /**
-     * Reads data from the adapter
-     *
-     * @param string     $key
-     * @param mixed|null $defaultValue
-     *
-     * @return mixed|null
-     * @throws StorageException
-     */
-    public function get(string $key, $defaultValue = null)
-    {
-        return $this->getUnserializedData(
-            $this->getAdapter()
-                 ->get($key),
-            $defaultValue
-        );
-    }
-
-    /**
      * Returns the already connected adapter or connects to the Memcached
      * server(s)
      *
