+++ src/Storage/Adapter/Redis.php
@@ -113,24 +113,6 @@ class Redis extends AbstractAdapter
     }
 
     /**
-     * Reads data from the adapter
-     *
-     * @param string $key
-     * @param null   $defaultValue
-     *
-     * @return mixed|null
-     * @throws StorageException
-     */
-    public function get(string $key, $defaultValue = null)
-    {
-        return $this->getUnserializedData(
-            $this->getAdapter()
-                 ->get($key),
-            $defaultValue
-        );
-    }
-
-    /**
      * Returns the already connected adapter or connects to the Redis
      * server(s)
      *
