+++ tests/_support/_generated/IntegrationTesterActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] c7820a4cba2f8bcc4cff811bcac98b0e
+<?php  //[STAMP] 1ed195bb381926e5dce09e9af7ea8dd8
 namespace _generated;
 
 // This class was automatically generated by build task
