+++ tests/_support/_generated/UnitTesterActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] a7aaf7f0444c44a99959e325f303f014
+<?php  //[STAMP] 1176e0e48adf80186675af3323b19fc0
 namespace _generated;
 
 // This class was automatically generated by build task
@@ -2466,7 +2466,7 @@ trait UnitTesterActions
         return $this->getScenario()->runStep(new \Codeception\Step\Action('callProtectedMethod', func_get_args()));
     }
 
-
+ 
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
@@ -2506,7 +2506,7 @@ trait UnitTesterActions
         return $this->getScenario()->runStep(new \Codeception\Step\Action('getProtectedProperty', func_get_args()));
     }
 
-
+ 
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
