+++ tests/integration/Session/Adapter/Redis/ReadWriteCest.php
@@ -14,41 +14,41 @@ declare(strict_types=1);
 namespace Phalcon\Tests\Integration\Session\Adapter\Redis;
 
 use IntegrationTester;
+use Phalcon\Session\Adapter\Redis;
 use Phalcon\Tests\Fixtures\Traits\DiTrait;
 
 use function uniqid;
 
-/**
- * Class WriteCest
- *
- * @package Phalcon\Tests\Integration\Session\Adapter\Redis
- */
-class WriteCest
+class ReadWriteCest
 {
     use DiTrait;
 
     /**
-     * Tests Phalcon\Session\Adapter\Redis :: write()
+     * Tests Phalcon\Session\Adapter\Redis :: read()
      *
      * @param IntegrationTester $I
      *
      * @author Phalcon Team <team@phalcon.io>
      * @since  2020-09-09
      */
-    public function sessionAdapterRedisWrite(IntegrationTester $I)
+    public function sessionAdapterRedisRead(IntegrationTester $I)
     {
-        $I->wantToTest('Session\Adapter\Redis - write()');
+        $I->wantToTest('Session\Adapter\Redis - read()');
 
+        /** @var Redis $adapter */
         $adapter = $this->newService('sessionRedis');
         $value   = uniqid();
+
         $adapter->write('test1', $value);
 
-        /**
-         * Serialize the value because the adapter does not have a serializer
-         */
-        $expected = serialize($value);
-        $actual   = $I->grabFromRedis('sess-reds-test1');
+//        $I->haveInRedis('string', 'sess-reds-test1', $value);
+
+        $expected = $value;
+        $actual   = $adapter->read('test1');
         $I->assertEquals($expected, $actual);
         $I->sendCommandToRedis('del', 'sess-reds-test1');
+
+        $actual = $adapter->read('test1');
+        $I->assertNotNull($actual);
     }
 }
